Vision Based Navigation for HAR/requirements.txt
adal==1.2.0
almath==2.9.1.32
asn1crypto==0.24.0
atomicwrites==1.2.1
attrs==18.2.0
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-cognitiveservices-knowledge-nspkg==3.0.0
azure-cognitiveservices-knowledge-qnamaker==0.1.0
azure-cognitiveservices-language-luis==0.1.0
azure-cognitiveservices-language-nspkg==3.0.0
azure-cognitiveservices-language-spellcheck==1.0.0
azure-cognitiveservices-language-textanalytics==0.1.0
azure-cognitiveservices-nspkg==3.0.0
azure-cognitiveservices-search-autosuggest==0.1.0
azure-cognitiveservices-search-customimagesearch==0.1.0
azure-cognitiveservices-search-customsearch==0.2.0
azure-cognitiveservices-search-entitysearch==1.0.0
azure-cognitiveservices-search-imagesearch==1.0.0
azure-cognitiveservices-search-newssearch==1.0.0
azure-cognitiveservices-search-nspkg==3.0.0
azure-cognitiveservices-search-videosearch==1.0.0
azure-cognitiveservices-search-visualsearch==0.2.0
azure-cognitiveservices-search-websearch==1.0.0
azure-cognitiveservices-vision-computervision==0.2.0
azure-cognitiveservices-vision-contentmoderator==0.2.0
azure-cognitiveservices-vision-customvision==0.4.0
azure-cognitiveservices-vision-face==0.1.0
azure-cognitiveservices-vision-nspkg==3.0.0
azure-common==1.1.16
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.5
azure-datalake-store==0.0.40
azure-devtools==1.1.1
azure-eventgrid==1.2.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-alertsmanagement==0.1.0
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-botservice==0.1.0
azure-mgmt-cdn==3.0.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.4.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.4.0
azure-mgmt-containerregistry==2.6.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-documentdb==0.1.3
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.2.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-hdinsight==0.2.0
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-kusto==0.2.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==0.1.0
azure-mgmt-network==2.5.0
azure-mgmt-notificationhubs==2.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.5.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.1.0
azure-mgmt-resourcegraph==0.1.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.0.0
azure-mgmt-security==0.1.0
azure-mgmt-servermanager==2.0.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-sqlvirtualmachine==0.2.0
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-sdk-tools==0.0.0
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.6
azure-storage-blob==1.4.0
azure-storage-common==1.4.0
azure-storage-file==1.4.0
azure-storage-nspkg==3.1.0
azure-storage-queue==1.4.0
bleach==3.1.0
cachetools==3.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
ConfigArgParse==0.13.0
configparser==3.5.0
contextlib2==0.5.5
coverage==5.0a4
cryptography==2.4.2
Deprecated==1.2.4
Django==1.11.18
docutils==0.14
entrypoints==0.3
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
get==2018.11.19
gitdb2==2.0.5
GitPython==2.1.11
google-api-core==1.7.0
google-auth==1.6.2
google-cloud==0.34.0
google-cloud-speech==0.36.1
google-cloud-vision==0.35.2
googleapis-common-protos==1.5.5
grpcio==1.17.1
idna==2.8
image==1.5.27
ipaddress==1.0.22
isodate==0.6.0
Jinja2==2.10
json-delta==2.0
keyring==17.1.1
MarkupSafe==1.1.0
mock==2.0.0
more-itertools==5.0.0
msrest==0.6.4
msrestazure==0.6.0
numpy==1.11.0
oauthlib==2.1.0
Owlready2==0.13
packaging==18.0
pathlib2==2.3.3
pbr==5.1.1
Pillow==5.4.1
pluggy==0.8.1
post==2018.11.20
protobuf==3.6.1
public==2018.11.20
py==1.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.3
pycparser==2.19
PyGithub==1.43.4
Pygments==2.3.1
PyJWT==1.7.1
pyOpenSSL==18.0.0
pyparsing==2.3.0
pytest==4.1.0
pytest-cov==2.6.1
python-dateutil==2.7.5
pytoml==0.1.20
pytz==2018.9
PyYAML==4.2b4
qi==1.6.3
query-string==2018.11.20
readme-renderer==24.0
request==2018.11.20
requests==2.21.0
requests-oauthlib==1.1.0
rsa==4.0
scandir==1.9.0
SecretStorage==2.3.1
six==1.12.0
smmap2==2.0.5
typing==3.6.6
urllib3==1.24.1
vcrpy==2.0.1
webencodings==0.5.1
wrapt==1.11.0
